github-pull-request-stats/requirements.txt
PyGithub==1.43.4
python-dotenv==0.10.1
